+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   turned on by default
 - Added a command line provider `config.NewCommandLineProvider()`, which can be used
   to pass configuration parameters through command line.
+- **[Breaking]** `uhttp module` now accepts `http.Handler` as part of module setup.
+  As part of refactor, RouteHandler is removed from the module registration.
 
 ## v1.0.0-beta3 (28 Mar 2017)